+++ api/dto/request.go
@@ -20,35 +20,57 @@ func (d DateTime) String() string {
 	return d.Time.UTC().Format("2006-01-02T15:04:05Z")
 }
 
+type pagination struct {
+	page     int
+	pageSize int
+}
+
+func NewPagination(page, size int) pagination {
+	return pagination{
+		page:     page,
+		pageSize: size,
+	}
+}
+
+// AppendToQuery decorates the URL with pagination parameters
+func (p pagination) AppendToQuery(u url.URL) url.URL {
+	v := u.Query()
+
+	if p.page != 0 {
+		v.Add("page", strconv.Itoa(p.page))
+	}
+	if p.pageSize != 0 {
+		v.Add("page-size", strconv.Itoa(p.pageSize))
+	}
+
+	u.RawQuery = v.Encode()
+
+	return u
+}
+
 // TimeEntryStartEndRequest to get entries by range
 type TimeEntryStartEndRequest struct {
 	Start    DateTime
 	End      DateTime
 	Hydrated *bool
 
-	page     int
-	pageSize int
+	Pagination pagination
 }
 
+// WithPagination add pagination to the TimeEntryStartEndRequest
 func (r TimeEntryStartEndRequest) WithPagination(page, size int) TimeEntryStartEndRequest {
-	r.page = page
-	r.pageSize = size
+	r.Pagination = NewPagination(page, size)
 	return r
 }
 
+// AppendToQuery decorates the URL with the query string needed for this Request
 func (r TimeEntryStartEndRequest) AppendToQuery(u url.URL) url.URL {
+	u = r.Pagination.AppendToQuery(u)
 	v := u.Query()
 	v.Add("start", r.Start.String())
 	v.Add("end", r.End.String())
 	if r.Hydrated != nil && *r.Hydrated {
 		v.Add("hydrated", "true")
-	}
-
-	if r.page != 0 {
-		v.Add("page", strconv.Itoa(r.page))
-	}
-	if r.pageSize != 0 {
-		v.Add("page-size", strconv.Itoa(r.pageSize))
 	}
 
 	u.RawQuery = v.Encode()
@@ -81,4 +103,26 @@ type UpdateTimeEntryRequest struct {
 	ProjectID   string    `json:"projectId,omitempty"`
 	TaskID      string    `json:"taskId,omitempty"`
 	TagIDs      []string  `json:"tagIds,omitempty"`
+}
+
+type GetProjectRequest struct {
+	Name     string
+	Archived bool
+
+	Pagination pagination
+}
+
+// AppendToQuery decorates the URL with the query string needed for this Request
+func (r GetProjectRequest) AppendToQuery(u url.URL) url.URL {
+	u = r.Pagination.AppendToQuery(u)
+
+	v := u.Query()
+	v.Add("name", r.Name)
+	if r.Archived {
+		v.Add("archived", "true")
+	}
+
+	u.RawQuery = v.Encode()
+
+	return u
 }